requirements.txt
flask
flask-cors
requests
python-dotenv
flask==2.2.2
flask-cors==3.1.1
requests==2.28.1
python-dotenv==0.21.0
gunicorn==20.1.0  # Added gunicorn for production use on Vercel
